rearranged Module Hirarchy. MakeIMG -> Dither -> Colorsplicer.
